refactor: R4b — models/traits/events/jobs at ZERO phpstan baseline entries
The long tail: natively-typed scope docblocks, assign-then-return for
the whereRaw/whereNull/whereDate generic-erasure trap, typed role/
permission APIs on User/Role/Permission/Team, relation generics for the
stragglers, date-cast @property tightening (Carbon|null, not
Carbon|string), typed event props (MachineLocationUpdated/MachineOffline
shapes), job payload shapes, and real null-safety guards in the
monitoring jobs (team/machine/timestamp/manufacturer_id/file-read
failure paths). One honest reversal: RouteSpeedMonitoringJob reads raw
DB::table rows — typing its param as MachineMetric was wrong; it is
documented stdClass. phpstan ledger 1032→890 with in-scope entries at
ZERO; psalm in-scope occurrences 1843→1571 as a side effect. Suite 601
green throughout.

+++ app/Jobs/AlertGenerationJob.php
@@ -77,6 +77,8 @@ class AlertGenerationJob implements ShouldQueue
 
     /**
      * Generate alerts for a specific machine based on its current state.
+     *
+     * @return list<array<string, mixed>>
      */
     private function generateAlertsForMachine(Machine $machine): array
     {
@@ -229,6 +231,8 @@ class AlertGenerationJob implements ShouldQueue
 
     /**
      * Create or update an alert in the database and broadcast it.
+     *
+     * @param  array<string, mixed>  $alertData
      */
     private function createOrUpdateAlert(Machine $machine, array $alertData): Alert
     {
@@ -284,7 +288,7 @@ class AlertGenerationJob implements ShouldQueue
     /**
      * Calculate hours until next maintenance based on machine type and hours.
      */
-    private function getHoursUntilMaintenance(Machine $machine): ?float
+    private function getHoursUntilMaintenance(Machine $machine): ?int
     {
         // Maintenance schedules based on machine type
         $maintenanceSchedule = [
